Share phenogram with invokers of main() in cli.cxx
First quick and dirty effort to ensure that both FlowJoBridge and AutoGate users can see the phenogram webpage produced in response to their request for EPP gating.
This first quick and dirty (33 minutes programming) effort leaves the invoker of main()  responsible for the file locations and correctness of phenogram.html and template.html.  When and IF more command line invocation environments emerge the work described in the TODO comment can then be undertaken.